UI/dominant-colors/requirements.txt
email-validator==1.1.3
Flask==2.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
Werkzeug==2.1.1
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
